threads: TODO 2nd threads test, still refcount issues RT#22760
In order to fix shared sets, we will need to change to global lock to check
only the SV set object for shared magic, and lock only with such a SV. This will need
a small non-public API change. Also the refcount and HEK deletion issue needs to be fixed.